[refactor/207-course-card] Extract useCourseCardTermState from CourseCard
Part of #207: move term, swap, and absence derivations into a hook
without behavior changes.
